Sequence locktime flags converted to long

+++ core/src/main/java/org/bitcoinj/core/Transaction.java
@@ -195,16 +195,16 @@ public class Transaction extends ChildMessage {
     /* Below flags apply in the context of BIP 68 */
     /* If this flag set, CTxIn::nSequence is NOT interpreted as a
      * relative lock-time. */
-    public static final int SEQUENCE_LOCKTIME_DISABLE_FLAG = 1 << 31;
+    public static final long SEQUENCE_LOCKTIME_DISABLE_FLAG = 1L << 31;
     
     /* If CTxIn::nSequence encodes a relative lock-time and this flag
      * is set, the relative lock-time has units of 512 seconds,
      * otherwise it specifies blocks with a granularity of 1. */
-    public static final int SEQUENCE_LOCKTIME_TYPE_FLAG = 1 << 22;
+    public static final long SEQUENCE_LOCKTIME_TYPE_FLAG = 1L << 22;
 
     /* If CTxIn::nSequence encodes a relative lock-time, this mask is
      * applied to extract that lock-time from the sequence field. */
-    public static final int SEQUENCE_LOCKTIME_MASK = 0x0000ffff;
+    public static final long SEQUENCE_LOCKTIME_MASK = 0x0000ffff;
     
     public Transaction(NetworkParameters params) {
         super(params);
